+++ app/controllers/stakeholder-controller.js
@@ -3,31 +3,6 @@ const { Readable } = require("stream");
 const stringify = require("csv-stringify");
 
 const search = (req, res) => {
-  let categoryIds = req.query.categoryIds;
-  if (!req.query.latitude || !req.query.longitude) {
-    res
-      .status(404)
-      .json("Bad request: needs latitude and longitude parameters");
-  }
-  if (!categoryIds) {
-    // If no filter, just use active categories.
-    categoryIds = ["1", "3", "8", "9", "10", "11", "12"];
-  } else if (typeof categoryIds == "string") {
-    categoryIds = [categoryIds];
-  }
-  const params = { ...req.query, categoryIds };
-  stakeholderService
-    .search(params)
-    .then((resp) => {
-      res.send(resp);
-    })
-    .catch((err) => {
-      console.log(err);
-      res.status("404").json({ error: err.toString() });
-    });
-};
-
-const searchDashboard = (req, res) => {
   if (req.distance && (!req.latitude || !req.longitude)) {
     res
       .status(404)
@@ -42,7 +17,7 @@ const searchDashboard = (req, res) => {
   }
   const params = { ...req.query, categoryIds };
   stakeholderService
-    .searchDashboard(params)
+    .search(params)
     .then((resp) => {
       res.send(resp);
     })
@@ -210,9 +185,8 @@ const claim = (req, res) => {
 
 module.exports = {
   search,
-  searchDashboard,
-  csv,
   getById,
+  csv,
   post,
   put,
   remove,
