+++ app/routes/index.js
@@ -8,6 +8,7 @@ const suggestionRouter = require("./suggestion-router");
 
 const faqRouter = require("./faq-router");
 const stakeholderRouter = require("./stakeholder-router");
+const stakeholderBestRouter = require("./stakeholder-best-router");
 const importRouter = require("./import-router");
 const loadRouter = require("./load-router");
 const esriRouter = require("./esri-router");
@@ -15,6 +16,7 @@ const esriRouter = require("./esri-router");
 module.exports = router;
 
 router.use("/api/stakeholders", stakeholderRouter);
+router.use("/api/stakeholderbests", stakeholderBestRouter);
 router.use("/api/tenants", tenantRouter);
 router.use("/api/accounts", accountRouter);
 router.use("/api/categories", categoryRouter);
