+++ client/src/hooks/useOrganizations.js
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import * as stakeholderService from "../../services/stakeholder-service";
+import * as stakeholderService from "../services/stakeholder-service";
 
 export const useOrganizations = () => {
   const [state, setState] = useState({
@@ -9,43 +9,6 @@ export const useOrganizations = () => {
   });
 
   const search = async ({
-    name,
-    latitude,
-    longitude,
-    radius,
-    categoryIds,
-    isInactive,
-    verificationStatusId,
-  }) => {
-    if (!latitude || !longitude) {
-      setState({ data: null, loading: false, error: true });
-      const msg =
-        "Call to search function missing latitude and/or longitude parameters";
-      console.error(msg);
-      return Promise.reject(msg);
-    }
-    //if (!categoryIds || categoryIds.length === 0) return;
-    try {
-      setState({ data: null, loading: true, error: false });
-      const stakeholders = await stakeholderService.search({
-        name,
-        categoryIds,
-        latitude,
-        longitude,
-        distance: radius,
-        isInactive,
-        verificationStatusId,
-      });
-      setState({ data: stakeholders, loading: false, error: false });
-      return stakeholders;
-    } catch (err) {
-      setState({ data: null, loading: false, error: true });
-      console.error(err);
-      return Promise.reject(err);
-    }
-  };
-
-  const searchDashboard = async ({
     tenantId,
     name,
     latitude,
@@ -68,7 +31,7 @@ export const useOrganizations = () => {
   }) => {
     try {
       setState({ data: null, loading: true, error: false });
-      const stakeholders = await stakeholderService.searchDashboard({
+      const stakeholders = await stakeholderService.search({
         tenantId,
         name,
         latitude,
@@ -98,5 +61,5 @@ export const useOrganizations = () => {
     }
   };
 
-  return { ...state, search, searchDashboard };
+  return { ...state, search };
 };
